+++ projects/core/src/lib/ng-progress-ref.ts
@@ -1,28 +1,36 @@
 import { NgProgressState, NgProgressConfig } from './ng-progress.interface';
-import { Observable, Subject, BehaviorSubject, timer, of, combineLatest } from 'rxjs';
+import { Observable, Subject, BehaviorSubject, timer, of, combineLatest, SubscriptionLike, Subscription } from 'rxjs';
 import { tap, map, skip, delay, filter, debounce, switchMap, distinctUntilChanged } from 'rxjs/operators';
 
 export class NgProgressRef {
 
-  private _state: NgProgressState = {active: false, value: 0};
-  private _config: NgProgressConfig;
-
-  /** Stream that increments and updates progress state */
-  private _trickling$ = new Subject();
-
   /** Stream that emits when progress state is changed */
-  state$ = new BehaviorSubject<NgProgressState>(this._state);
+  private readonly _state: BehaviorSubject<NgProgressState>;
+  state: Observable<NgProgressState>;
 
   /** Stream that emits when config is changed */
-  config$ = new Subject<NgProgressConfig>();
+  private readonly _config: BehaviorSubject<NgProgressConfig>;
+  config: Observable<NgProgressState>;
 
+  /** Stream that increments and updates progress state */
+  private readonly _trickling = new Subject();
+
+  /** Stream that combines "_trickling" and "config" streams */
+  private readonly _worker: SubscriptionLike = Subscription.EMPTY;
+
+  /** Get current progress state */
+  private get currState(): NgProgressState {
+    return this._state.value;
+  }
+
+  /** Check if progress has started */
   get isStarted(): boolean {
-    return this._state.active;
+    return this.currState.active;
   }
 
   /** Progress start event */
   get started(): Observable<boolean> {
-    return this.state$.pipe(
+    return this._state.pipe(
       map((state: NgProgressState) => state.active),
       distinctUntilChanged(),
       filter(active => active)
@@ -31,7 +39,7 @@ export class NgProgressRef {
 
   /** Progress ended event */
   get completed(): Observable<boolean> {
-    return this.state$.pipe(
+    return this._state.pipe(
       map((state: NgProgressState) => state.active),
       distinctUntilChanged(),
       filter(active => !active),
@@ -39,86 +47,118 @@ export class NgProgressRef {
     );
   }
 
-  constructor(customConfig: NgProgressConfig) {
+  constructor(customConfig: NgProgressConfig, private deleteInstance: Function) {
+    this._state = new BehaviorSubject<NgProgressState>({active: false, value: 0});
+    this._config = new BehaviorSubject<NgProgressConfig>(customConfig);
+    this.state = this._state.asObservable();
+    this.config = this._state.asObservable();
 
-    combineLatest(this._trickling$, this.config$).pipe(
-      debounce(([start, config]: [boolean, NgProgressConfig]) => timer(start ? this._config.debounceTime : 0)),
-      switchMap(([start, config]: [boolean, NgProgressConfig]) => start ? this._trickling(config) : this._complete(config))
+    this._worker = combineLatest(this._trickling, this._config).pipe(
+      debounce(([start, config]: [boolean, NgProgressConfig]) => timer(start ? config.debounceTime : 0)),
+      switchMap(([start, config]: [boolean, NgProgressConfig]) => start ? this.onTrickling(config) : this.onComplete(config))
     ).subscribe();
-
-    this.setConfig(customConfig);
   }
 
+  /**
+   * Start the progress
+   */
   start() {
-    this._trickling$.next(true);
+    this._trickling.next(true);
   }
 
+  /**
+   * Complete the progress
+   */
   complete() {
-    this._trickling$.next(false);
+    this._trickling.next(false);
   }
 
+  /**
+   * Increment the progress
+   * @param amount
+   */
   inc(amount?: number) {
-    const n = this._state.value;
+    const n = this.currState.value;
     if (!this.isStarted) {
       this.start();
     } else {
       if (typeof amount !== 'number') {
-        amount = this._config.trickleFunc(n);
+        amount = this._config.value.trickleFunc(n);
       }
       this.set(n + amount);
     }
   }
 
+  /**
+   * Set the progress
+   * @param n
+   */
   set(n: number) {
-    this._setState({value: this._clamp(n), active: true});
-  }
-
-  setConfig(config: NgProgressConfig) {
-    this._config = {...this._config, ...config};
-    this.config$.next(this._config);
+    this.setState({value: this.clamp(n), active: true});
   }
 
   /**
-   * Meant to be used internally and not by user directly
-   * Users should use NgProgressManager.destroy(id) instead
+   * Set config
+   * @param config
+   */
+  setConfig(config: NgProgressConfig) {
+    this._config.next({...this._config.value, ...config});
+  }
+
+  /**
+   * Destroy progress reference
    */
   destroy() {
-    this._trickling$.complete();
-    this.state$.complete();
-    this.config$.complete();
+    this._worker.unsubscribe();
+    this._trickling.complete();
+    this._state.complete();
+    this._config.complete();
+    this.deleteInstance();
   }
 
-  private _setState(state: NgProgressState) {
-    this._state = {...this._state, ...state};
-    this.state$.next(this._state);
+  /**
+   * Set progress state
+   * @param state
+   */
+  private setState(state: NgProgressState) {
+    this._state.next({...this.currState, ...state});
   }
 
-  /** Clamps a value to be between min and max */
-  private _clamp(n): number {
-    return Math.max(this._config.min, Math.min(this._config.max, n));
+  /**
+   * Clamps a value to be between min and max
+   * @param n
+   */
+  private clamp(n: number): number {
+    return Math.max(this._config.value.min, Math.min(this._config.value.max, n));
   }
 
-  /** Keeps incrementing the progress */
-  private _trickling(config: NgProgressConfig) {
+  /**
+   * Keeps incrementing the progress
+   * @param config
+   */
+  private onTrickling(config: NgProgressConfig): Observable<number> {
     if (!this.isStarted) {
-      this.set(this._config.min);
+      this.set(this._config.value.min);
     }
     return timer(0, config.trickleSpeed).pipe(tap(() => this.inc()));
   }
 
-  /** Completes then resets the progress */
-  private _complete(config: NgProgressConfig) {
+  /**
+   * Completes then resets the progress
+   * @param config
+   */
+  private onComplete(config: NgProgressConfig): Observable<any> {
     return !this.isStarted ? of({}) : of({}).pipe(
       // Completes the progress
-      tap(() => this._setState({value: 100})),
+      tap(() => this.setState({value: 100})),
 
       // Hides the progress bar after a tiny delay
       delay(config.speed * 1.7),
-      tap(() => this._setState({active: false})),
+      tap(() => this.setState({active: false})),
 
       // Resets the progress state
       delay(config.speed),
-      tap(() => this._setState({value: 0}))
+      tap(() => this.setState({value: 0}))
     );
   }
 }