+++ projects/router/src/lib/ng-progress-router.interface.ts
@@ -1,8 +1,11 @@
-import { InjectionToken } from '@angular/core';
+import { InjectionToken, Type } from '@angular/core';
+import { RouterEvent } from '@angular/router';
 
 export interface NgProgressRouterConfig {
   id?: string;
   delay?: number;
+  startEvents?: Type<RouterEvent>[];
+  completeEvents?: Type<RouterEvent>[];
 }
 
 export const CONFIG = new InjectionToken<NgProgressRouterConfig>('config');